Reverted changes

+++ sst/sst.py
@@ -66,7 +66,7 @@ def main():
     dev_data = reader.read('https://s3-us-west-2.amazonaws.com/allennlp/datasets/sst/dev.txt')
     reader = StanfordSentimentTreeBankDatasetReader(granularity="2-class",
                                                     token_indexers={"tokens": single_id_indexer})
-    test_data = reader.read('https://s3-us-west-2.amazonaws.com/allennlp/datasets/sst/dev.txt')
+    test_data = reader.read('https://s3-us-west-2.amazonaws.com/allennlp/datasets/sst/test.txt')
 
     vocab = Vocabulary.from_instances(train_data)
 
@@ -111,28 +111,28 @@ def main():
     vocab_path = "tmp\\" + EMBEDDING_TYPE + "_" + "vocab"
 
     # if the model already exists (its been trained), load the pre-trained weights and vocabulary
-    # if os.path.isfile(model_path):
-    #     vocab = Vocabulary.from_files(vocab_path)
-    #     model = LstmClassifier(word_embeddings, encoder, vocab).cuda()
-    #     with open(model_path, 'rb') as f:
-    #         model.load_state_dict(torch.load(f))
+    if os.path.isfile(model_path):
+        vocab = Vocabulary.from_files(vocab_path)
+        model = LstmClassifier(word_embeddings, encoder, vocab).cuda()
+        with open(model_path, 'rb') as f:
+            model.load_state_dict(torch.load(f))
     # otherwise train model from scratch and save its weights
-    # else:
-    iterator = BucketIterator(batch_size=32, sorting_keys=[("tokens", "num_tokens")])
-    iterator.index_with(vocab)
-    optimizer = optim.Adam(model.parameters())
-    trainer = Trainer(model=model,
-                        optimizer=optimizer,
-                        iterator=iterator,
-                        train_dataset=train_data,
-                        validation_dataset=dev_data,
-                        num_epochs=5,
-                        patience=1,
-                        cuda_device=0)
-    trainer.train()
-    with open(model_path, 'wb') as f:
-        torch.save(model.state_dict(), f)
-    vocab.save_to_files(vocab_path)
+    else:
+        iterator = BucketIterator(batch_size=32, sorting_keys=[("tokens", "num_tokens")])
+        iterator.index_with(vocab)
+        optimizer = optim.Adam(model.parameters())
+        trainer = Trainer(model=model,
+                            optimizer=optimizer,
+                            iterator=iterator,
+                            train_dataset=train_data,
+                            validation_dataset=dev_data,
+                            num_epochs=5,
+                            patience=1,
+                            cuda_device=0)
+        trainer.train()
+        with open(model_path, 'wb') as f:
+            torch.save(model.state_dict(), f)
+        vocab.save_to_files(vocab_path)
     model.train().cuda() # rnn cannot do backwards in train mode
 
     # Register a gradient hook on the embeddings. This saves the gradient w.r.t. the word embeddings.
@@ -162,6 +162,7 @@ def main():
 
     # Get accuracy for triggers we want to test
     if test_triggers is not None:
+        utils.get_accuracy(model, targeted_dev_data, vocab, trigger_token_ids=None)
         utils.get_accuracy(model, targeted_test_data, vocab, trigger_token_ids=test_triggers)
         return
 
@@ -171,11 +172,12 @@ def main():
 
     negative_words_file = "negative_words.txt"
     positive_words_file = "positive_words.txt"
+    # neutral_words_file = "vader_neutral_words.txt"
 
     # Initialize the set of token ids to filter out from candidates
     token_id_to_filter = {}
 
-    # Read in negative words and add them to the set
+    # # Read in negative words and add them to the set
     with open(negative_words_file) as f:
         for line in f:
             token_id_to_filter[vocab.get_token_index(line.strip())] = True
@@ -184,6 +186,15 @@ def main():
     with open(positive_words_file) as f:
         for line in f:
             token_id_to_filter[vocab.get_token_index(line.strip())] = True
+
+    # Read in neutral words; if a token in the vocab isn't neutral, add to set
+    # neutral_set = set()
+    # with open(neutral_words_file) as f:
+    #     for line in f:
+    #         neutral_set.add(vocab.get_token_index(line.strip()))
+    # for token_idx in range(vocab.get_vocab_size()):
+    #     if token_idx not in neutral_set:
+    #         token_id_to_filter[token_idx] = True
 
     # initialize triggers which are concatenated to the input
     num_trigger_tokens = 3
